Intent: Consolidate structured IO helpers under tnfr.utils to clarify module topology Operators involved: [Self-organization, Coherence] Affected invariants: [#7, #8] Key changes: - Moved json serialization and structured file helpers into new tnfr.utils.io module - Updated tnfr.io to delegate to the consolidated implementation while exposing the same API - Adjusted utils exports, cache helpers, production code, and tests to import from the new entry point - Added lazy import in cache utilities to avoid cyclic dependencies during module bootstrap Expected risks/dissonances: Potential regressions from import-order changes mitigated by dedicated IO test suite runs Metrics: C(t) stable; Si unchanged; νf unaffected; phase relationships preserved Equivalence map: tnfr.io.* → tnfr.utils.io.*
