+++ tests/unit/structural/test_missing_dependency_error_cache.py
@@ -1,4 +1,4 @@
-from tnfr.io import _MISSING_TOML_ERROR, _MISSING_YAML_ERROR
+from tnfr.utils.io import _MISSING_TOML_ERROR, _MISSING_YAML_ERROR
 
 
 def test_missing_dependency_error_cached() -> None:
